Fix casts
Commit migrated from https://github.com/dotnet/coreclr/commit/13001f90aa71a7a0ec4039922f29087680b44ec3

+++ src/coreclr/src/gc/gcpriv.h
@@ -37,7 +37,7 @@ inline void FATAL_GC_ERROR()
     GCToOSInterface::DebugBreak();
 #endif // DACCESS_COMPILE
     _ASSERTE(!"Fatal Error in GC.");
-    GCToEEInterface::HandleFatalError(COR_E_EXECUTIONENGINE);
+    GCToEEInterface::HandleFatalError((unsigned int)COR_E_EXECUTIONENGINE);
 }
 
 #ifdef _MSC_VER
